\name{varbvs-package}
\alias{varbvs-package}
\alias{varbvs}
\docType{package}
\title{Variational inference for Bayesian variable selection}

\description{The \pkg{varbvs} pacckage implements the variational
inference procedure for Bayesian variable selection, as described in the
"Scalable variational inference for Bayesian variable selection in
regression, and its accuracy in genetic association studies" (Bayesian
Analysis 7, March 2012, pages 73-108). This software has been used to
implement Bayesian variable selection for large problems with over a
million variables and thousands of samples.}

\details{
\tabular{ll}{
Package: \tab varbvs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-22\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
