\name{varbvs-package}
\alias{varbvs-package}
\alias{varbvs}
\docType{package}
\title{Variational inference for Bayesian variable selection}

\description{The \pkg{varbvs} package implements the variational
inference procedure for Bayesian variable selection, as described in the
"Scalable variational inference for Bayesian variable selection in
regression, and its accuracy in genetic association studies" (Bayesian
Analysis 7, March 2012, pages 73-108). This software has been used to
implement Bayesian variable selection for large problems with over a
million variables and thousands of samples.}

\details{

TO DO: point to the high-level functions, and give an overview of what
they do. This should be done with links to the functions, as in
\code{\link{varbvsoptimize}}.
  
To see an example of how the variational inference procedure can be used
to compute posterior inclusion probabilities and the posterior
distribution of the hyperparameters in a linear regression setting, run
\code{demo(example1)}.
  
To see more details about the \pkg{varbvs} package, including the
license and the list of user-level functions available in this package,
use \code{help(package="varbvs")}.}

\author{Peter Carbonetto <pcarbo@uchicago.edu>}}

\references{Peter Carbonetto and Matthew Stephens (2012). Scalable
variational inference for Bayesian variable selection in regression, and
its accuracy in genetic association studies. Bayesian Analysis 7(1),
pages 73-108.}


